Addressing PR code
